% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_covid}
\alias{sce_covid}
\title{COVID-19 data}
\format{
sce_covid a SingleCellExperiment object
}
\usage{
data(sce_covid)
}
\description{
This is a filtered single-cell dataset of COVID-19 patient from the
Stephenson et al 2021 dataset.
}
\references{
Stephenson, E., Reynolds, G., Botting, R.A. et al. Single-cell multi-omics analysis of the immune response in COVID-19. Nat Med 27, 904–916 (2021). https://doi.org/10.1038/s41591-021-01329-2
}
\keyword{datasets}
